# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Backcast API
  version: "1.0.0"
  description: |
    **Backcast** — an AI SRE Incident Commander with agentic memory on CockroachDB + AWS.

    The deployed surface spans three Lambdas, each on its own host:

    | Component  | Host (server)                                   | Purpose                                            |
    |------------|-------------------------------------------------|----------------------------------------------------|
    | Ingest     | API Gateway `…/prod` **or** the ingest Fn URL   | Turn an alert webhook into an incident (HMAC-auth) |
    | Commander  | Commander Function URL                          | Run one Incident Commander agent turn (Nova Pro)   |
    | Web / API  | Webapp Function URL                             | Interactive demo UI + the live memory panels       |

    ### Authentication
    The **ingest** endpoint is protected by an **HMAC-SHA256 webhook signature** (see the
    `webhookHmac` security scheme). The commander and webapp panels are open in the demo
    deployment (guard them the same way for production).

    HMAC scheme: sign the bytes `"<x-backcast-timestamp>." + <raw body>` with the shared
    secret (Secrets Manager id `backcast/webhook-secret`); send
    `x-backcast-signature: sha256=<hexdigest>` and `x-backcast-timestamp: <unix seconds>`.
    Requests older than 300s are rejected (replay protection).
  license:
    name: Apache-2.0
servers:
...
paths:
...
  /openapi.yaml:
    get:
      tags: [meta]
      operationId: openapiSpec
      summary: This OpenAPI document (machine-readable)
      description: Interactive Swagger UI for the webapp routes is also served at `GET /docs`.
      responses:
        '200':
          description: OpenAPI 3.0 YAML
          content:
            application/yaml:
              schema: { type: string }
        '404':
          description: Spec not bundled in this deployment image
...
  /api/counterfactual:
    post:
      tags: [demo]
      operationId: runCounterfactual
      summary: Rewind → fork → simulate → compare → learn
      description: |
        Forks a chosen library scenario at a captured HLC, simulates every candidate remediation
        deterministically, ranks them, and returns the **decision regret** of the action actually
        taken versus the best branch.
      parameters:
        - $ref: '#/components/parameters/Org'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                scenario_key: { type: string, example: db_pool_exhaustion }
                actual_remediation: { type: string, example: restart-service }
      responses:
        '200':
          description: Counterfactual comparison
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CounterfactualResult' }
  /api/simulate:
    post:
      tags: [demo]
      operationId: runSimulate
      summary: Build-your-own counterfactual (custom scenario)
      description: |
        Define a completely custom incident — a hidden true cause plus remediations with their own
        effects (fixes/relieves), risk, cost, and recovery time — and get the same live decision-regret
        ranking. Computed by the deterministic model; no database write.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [true_cause, remediations]
              properties:
                true_cause: { type: string, example: a bad config push disabled request caching }
                description: { type: string }
                actual:
                  type: array
                  items: { type: string }
                remediations:
                  type: object
                  additionalProperties:
                    type: object
                    properties:
                      fixes: { type: boolean }
                      relieves: { type: boolean }
                      recovery_seconds: { type: number }
                      risk: { type: number }
                      cost: { type: number }
      responses:
        '200':
          description: Custom counterfactual comparison
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CounterfactualResult' }
        '400':
          description: true_cause and at least one remediation are required
  /api/incident:
    post:
      tags: [demo]
      operationId: runIncident
      summary: Belief revision + temporal no-leak reconstruction
      description: |
        Records early evidence, sets initial beliefs, captures an HLC, then adds *later* deploy
        evidence and revises beliefs. Reconstructs the incident **as of** the earlier HLC to prove
        the future evidence does not leak backwards (`no_leak: true`), and returns the belief flip.
      parameters:
        - $ref: '#/components/parameters/Org'
      responses:
        '200':
          description: Temporal reconstruction + belief history
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IncidentResult' }
  /api/race:
    post:
      tags: [demo]
      operationId: runRace
      summary: Action-lease concurrency + fencing + crash-safety
      description: |
        N workers race to claim one action lease (exactly one wins). Then a holder "crashes",
        a second worker takes over with a bumped fencing generation, and the revived stale worker
        is **fenced out** — the external effect executes exactly once.
      parameters:
        - $ref: '#/components/parameters/Org'
        - name: workers
          in: query
          schema: { type: integer, default: 20, minimum: 2, maximum: 50 }
      responses:
        '200':
          description: Race + fencing outcome
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RaceResult' }
  /api/agent:
    post:
      tags: [demo]
      operationId: runAgent
      summary: Live Incident Commander turn (Amazon Nova Pro)
      description: |
        Creates an incident from a free-form alert signal and runs one real agent turn in a Bedrock
        tool-use loop (recall → observe → assess beliefs → claim a fenced action lease → resolve),
        returning the tool trace, the claimed action, the beliefs after the turn, and the summary.
      parameters:
        - $ref: '#/components/parameters/Org'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [signal]
              properties:
                signal: { type: string, example: "payments-api 5xx; DB pool saturated after deploy d-8842" }
                scenario_key: { type: string, description: optional seed scenario (makes it replay-eligible) }
      responses:
        '200':
          description: Agent turn result
          content:
            application/json:
              schema:
                type: object
                properties:
                  incident_id: { type: string }
                  steps: { type: integer }
                  tool_calls: { type: array, items: { type: string } }
                  claimed_action: { type: string, nullable: true }
                  summary: { type: string }
                  ledger_verified: { type: boolean }
                  beliefs:
                    type: array
                    items:
                      type: object
                      properties:
                        statement: { type: string }
                        confidence: { type: number }
        '400':
          description: signal is required
components:
  securitySchemes:
    webhookHmac:
      type: apiKey
      in: header
      name: x-backcast-signature
      description: |
        HMAC-SHA256 over `"<x-backcast-timestamp>." + <raw body>`, hex-encoded and prefixed
        `sha256=`. Requires the companion `x-backcast-timestamp` header (unix seconds, ≤300s old).
  parameters:
    Org:
      name: org
      in: query
      description: Tenant / organization id.
      schema: { type: string, default: demo }
...
  schemas:
...
    CounterfactualResult:
      type: object
      properties:
        incident_id: { type: string, format: uuid }
        scenario: { type: string, example: db_pool_exhaustion }
        forked_at_hlc: { type: string, description: HLC the branches were forked from. }
        best_label: { type: string, example: "fork:rollback-deploy" }
        decision_regret:
          type: number
          format: float
          description: best.score − actual.score (0 means the action taken was optimal).
          example: 1.24
        lesson: { type: string, description: Verified best remediation, promoted to procedural memory. }
        ledger_verified: { type: boolean }
        branches:
          type: array
          items: { $ref: '#/components/schemas/Branch' }
    Branch:
      type: object
      properties:
        label: { type: string, example: "fork:rollback-deploy" }
        remediations:
          type: array
          items: { type: string }
        score: { type: number, format: float }
        recovered: { type: boolean }
        recurred: { type: boolean }
        time_to_recovery_s: { type: number, nullable: true }
        risk: { type: number }
        cost: { type: number }
        unnecessary_actions: { type: integer }
        is_actual: { type: boolean }
        is_best: { type: boolean }
    IncidentResult:
      type: object
      properties:
        incident_id: { type: string, format: uuid }
        t1_hlc: { type: string }
        surge_hypothesis: { type: string }
        deploy_hypothesis: { type: string }
        no_leak:
          type: boolean
          description: True iff the deploy evidence added after t1 is invisible in the t1 reconstruction.
        at_t1: { $ref: '#/components/schemas/Snapshot' }
        now: { $ref: '#/components/schemas/Snapshot' }
        deploy_belief_history:
          type: array
          items:
            type: object
            properties:
              confidence: { type: number, format: float }
              current: { type: boolean }
        ledger_verified: { type: boolean }
    Snapshot:
      type: object
      properties:
        evidence:
          type: array
          items:
            type: object
            properties:
              kind: { type: string }
              content: { type: string }
        surge: { type: number, format: float, description: Confidence in the traffic-surge hypothesis. }
        deploy: { type: number, format: float, description: Confidence in the bad-deploy hypothesis. }
    RaceResult:
      type: object
      properties:
        workers: { type: integer, example: 20 }
        winners: { type: integer, description: Always 1 — exactly one worker wins the lease., example: 1 }
        crash_takeover_generation: { type: integer, description: Fencing generation after takeover., example: 2 }
        taker_completed: { type: boolean }
        revived_stale_worker_accepted:
          type: boolean
          description: False — the revived crashed worker is fenced out.
          example: false
        canonical_action_owners:
          type: integer
          description: >-
            Always 1 — exactly one logical owner holds the action at a time. This is the
            guarantee Backcast makes; it does not claim exactly-once external execution,
            because a database transaction cannot atomically commit an external side effect.
          example: 1
        simulated_effect_applications:
          type: integer
          description: >-
            Always 1 — an idempotency-key-guarded write, applied once across the crash and
            the revival. It is a stand-in for a remediation inside this same cluster, not an
            external side effect; this build never executes a real remediation.
          example: 1
...